blog/requirements.txt
beautifulsoup4==4.9.3
Django==2.2.13
djangorestframework==3.12.2
Markdown==3.3.3
pytz==2018.5
